requirements.txt
pytest==7.4.0
pytest-selenium==4.0.1
pytest-xdist==3.3.1
selenium==4.11.2
allure-pytest==2.13.2
pymysql==1.1.0
webdriver-manager==4.0.0
retrying==1.3.4